# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me": {
      "get": {
        "operationId": "UsersController_getProfile",
        "summary": "Get the current user's profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profile (password hash and verify token stripped)"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "patch": {
        "operationId": "UsersController_updateProfile",
        "summary": "Update the current user's profile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated"
          },
          "400": {
            "description": "Validation failed on the request body"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "put": {
        "operationId": "UsersController_updateProfilePut",
        "summary": "Update the current user's profile (PUT alias)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "UsersController_deleteAccount",
        "summary": "Delete (soft-delete) the current user's account",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Account soft-deleted; sessions revoked"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/users/profile": {
      "get": {
        "operationId": "UsersController_getProfileAlias",
        "summary": "Get the current user's profile (alias)",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/users/me/password": {
      "patch": {
        "operationId": "UsersController_changePassword",
        "summary": "Change the current password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed; existing sessions revoked"
          },
          "400": {
            "description": "New password same as current password"
          },
          "401": {
            "description": "Current password is incorrect"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/users/change-password": {
      "post": {
        "operationId": "UsersController_changePasswordPost",
        "summary": "Change the current password (POST alias)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/subscriptions/status": {
      "get": {
        "operationId": "SubscriptionsController_getStatus",
        "summary": "Get current subscription status",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Plan, status, and renewal/cancellation dates"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/subscriptions/me": {
      "get": {
        "operationId": "SubscriptionsController_getStatusMe",
        "summary": "Get current subscription status (alias)",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/subscriptions/usage": {
      "get": {
        "operationId": "SubscriptionsController_getUsage",
        "summary": "Get remaining requests for today",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Daily limit, requests used, and requests remaining"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/subscriptions/remaining": {
      "get": {
        "operationId": "SubscriptionsController_getRemaining",
        "summary": "Get remaining requests for today (alias)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Daily limit, requests used, and requests remaining"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/subscriptions/plan": {
      "patch": {
        "operationId": "SubscriptionsController_changePlan",
        "summary": "Upgrade or downgrade subscription plan",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plan changed; usage limits updated accordingly"
          },
          "400": {
            "description": "Validation failed (invalid plan value)"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/subscriptions/upgrade": {
      "post": {
        "operationId": "SubscriptionsController_upgrade",
        "summary": "Upgrade subscription to PREMIUM",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Upgraded to PREMIUM"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/subscriptions/downgrade": {
      "post": {
        "operationId": "SubscriptionsController_downgrade",
        "summary": "Downgrade subscription to FREE",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Downgraded to FREE"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/subscriptions/cancel": {
      "post": {
        "operationId": "SubscriptionsController_cancel",
        "summary": "Cancel the current subscription (reverts to Free at period end)",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Subscription marked as canceled"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "EchoGPT Backend API",
    "description": "Production-ready backend for the EchoGPT Chrome Extension: auth, subscriptions, AI provider management, chat, web search, and admin analytics.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Jane Doe"
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "ChangePlanDto": {
        "type": "object",
        "properties": {
          "plan": {
            "type": "string",
            "enum": [
              "FREE",
              "PREMIUM"
            ]
          }
        },
        "required": [
          "plan"
        ]
      },
...
    }
  }
}